worker/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
httpx==0.28.1
beautifulsoup4==4.12.3
torch==2.5.1
transformers==4.47.1
accelerate==1.2.1
numpy==2.2.1
pandas==2.2.3
scikit-learn==1.6.0
pydantic==2.10.4
python-dotenv==1.0.1
